improve encodeHex, and decodeHex
